feat: new icons and bold text for the current step
issue #1434
